Fix wrong pinout for 2nd I2C bus (#2)
The ESP32-S3 box 3 uses 2 I2C ports, one on GPIO 8/18 and the other on GPIO 40/41. The fix allows to use the second port that's connected to sensors like AHT30 and radar. 

Currently, even when declaring the I2C GPIO in the YAML file, this function changes the GPIO **after** ESPHome sets them which is wrong.